+++ src/popups/taglist.rs
@@ -216,9 +216,19 @@ impl Component for TagListPopup {
 					self.key_config.keys.page_down,
 				) {
 					self.move_selection(ScrollType::PageDown);
+				} else if key_match(
+					key,
+					self.key_config.keys.page_half_down,
+				) {
+					self.move_selection(ScrollType::PageHalfDown);
 				} else if key_match(key, self.key_config.keys.page_up)
 				{
 					self.move_selection(ScrollType::PageUp);
+				} else if key_match(
+					key,
+					self.key_config.keys.page_half_up,
+				) {
+					self.move_selection(ScrollType::PageHalfUp);
 				} else if key_match(
 					key,
 					self.key_config.keys.move_right,
@@ -400,10 +410,16 @@ impl TagListPopup {
 			ScrollType::PageUp => old_selection.saturating_sub(
 				self.current_height.get().saturating_sub(1),
 			),
+			ScrollType::PageHalfUp => old_selection.saturating_sub(
+				self.current_height.get().div_ceil(2),
+			),
 			ScrollType::PageDown => old_selection
 				.saturating_add(
 					self.current_height.get().saturating_sub(1),
 				)
+				.min(max_selection),
+			ScrollType::PageHalfDown => old_selection
+				.saturating_add(self.current_height.get().div_ceil(2))
 				.min(max_selection),
 		};